requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
jiter==0.5.0
multidict==6.1.0
numpy==1.26.4
openai==1.51.0
passlib==1.7.4
phonenumbers==8.13.47
pyarrow==17.0.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pylance==0.18.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tqdm==4.66.5
twilio==9.3.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
wsproto==1.2.0
yarl==1.13.1
